+++ test/libsolidity/syntaxTests/functionTypes/function_type_parameters_with_names.sol
@@ -1,0 +1,5 @@
+contract C {
+    function(uint a) f;
+}
+// ----
+// SyntaxError: (26-32): Arguments in function types may not be named.